Add type column to overwatch page
Don't merge this yet, it blows up! I'm getting this error:

```
[error] an exception was raised:
** (Protocol.UndefinedError) protocol Enumerable not implemented for #Ecto.Association.NotLoaded<association :reports is not loaded> of type Ecto.Association.NotLoaded (a struct)
```

So I understand that I would need to pre-load reports for each
`report_group` in `lib/teiserver_web/live/moderation/overwatch/index.ex`
for this to work. I'm not sure how to do that, though... I gather that you're
pre-loading the report groups here:

```
    socket = socket
    |> assign(:site_menu_active, "moderation")
    |> assign(:view_colour, Teiserver.Moderation.colour())
    |> assign(:outstanding_report_groups, 0)
    |> assign(:report_groups, nil)                 <----- this?
    |> default_filters(params)
    |> add_breadcrumb(name: "Moderation", url: ~p"/moderation")
    |> add_breadcrumb(name: "Overwatch", url: ~p"/moderation/overwatch")
```

But I'm not sure how to extend that to also load something "belonging"
to those report groups.