Fix broken link on reports index page
The green user icon was pointing to moderation/report/<user_id> or some such, this got removed a while back in the overwatch refactor and would cause Internal Server Error.

Lexon has, since my original branch was created, replaced that with a profile_overview_path link, but this page is actually more useful for moderation.