Skeleton autohost table
I'm almost sure teiserver is going to use a different table for
autohosts going forward with tachyon. This is cleaner than relying on
permissions of the user to differentiate between hosts and players.
This table is there only so that the client secrets for OAuth can
reference the correct table.