Add "$set-config-teaser" Host-command.
This command allows Hosts to specify a string that can be used
to quickly describe a lobby's currently-configured settings.

The teaser is currently used in the title of a lobby, similar to
the existing behavior but without using the "$rename" command. This
gives the benefit of not potentially overriding a custom player-set
lobby name.