Make web-based lobby chat read-only for non-Admin/Moderator users.
Currently, any non-Muted user is allowed to read and send lobby
chat messages via the website.

However, the use-case for this feature is not very strong for the
typical player, and we occasionally see it abused by players who
have been kickbanned from the lobby (to flame anyone involved in
the kickban).

Therefore, this patch restricts web-based chat such that only users
with either the "Admin" or "Moderator" roles are able to send
messages.