Use same max length for `!welcome-message` and `$welcome-message`. (#407)
Teiserver enforces a maximum message length for the `SAYBATTLE` and
`SAYBATTLEEX` Spring messages. These messages are limited to 256
characters by default, but they are increased if the message begins
with certain prefixes.

Currently, when these messages begin with `$welcome-message`, the
limit is raised to 1024 characters.

This patch adds `!welcome-message` to the that special case check,
so that both versions of the command have the same limit of 1024.