Remove vacuum FULL from daily cleanup task
Only keep the daily vacuum analyze query. This one can be run
concurrently with other read/write operation.
The cost is that some deleted rows would still occupy space on disk, but
hopefully it won't cause logging and disconnection issue.

This is a bit of a stopgap measure until we figure out the true cause of
the morning disconnections.