Purge almost all mention of tachyon (#378)
There used to be two prototype versions of the protocol. One TLS based
and one Websocket based. None of them were fully functionning and since
the departure of Teifion from the project the protocol evolved more.

Having legacy code lingering around caused some confusion when trying to
understand the server.

Note that there may be additional cleanup required. A lot of the tachyon
system was based around Coordinator and Consul, and they may or may not
be tachyon logic in them. We'll figure that out later.

There is one piece of logic that could probably be reused almost as-is:
the function `login_from_token` (search the diffs). Swapping the token
code for the OAuth based one, the rest should be almost identical.