Make new login sessions replace existing sessions. (#422)
* Make new login sessions replace existing sessions.

Currently, if a session already exists for a user and a client
attempts to log in as the same user, the following happens:
- The login flood protection count is incremented
- The existing session is disconnected
- The new login attempt is rejected with instructions to try later

This patch changes that last point. Now, the following happens:
- The login flood protection count is incremented
- The existing session is disconnected
- The new login attempt proceeds as usual

The login attempt may still fail, for example if Teiserver's number
of active sessions is at capacity; however, for typical cases the
login will succeed.

Note: most of the usual login checks occur before any existing
sessions are disconnected (such as password verification). That is
unchanged by this patch.

* Swap order of delay/disconnect when replacing existing logins.