Allow chat-restricted users to run whitelisted commands.
This patch allows users with an active chat restriction to use
several common/necessary commands. Limitations are placed on both
the commands that are allowed, as well as on the parameters (if
any) allowed with each command:
- `!vote`, and the various yes/no/blank parameters for it
- `!endvote` and `!ev`
- `!help`
- `!balance` (or a vote for it)
- `!start` (or a vote for it)
- `!forcestart` (or a vote for it)
- `!stop` (or a vote for it)
- `!resign` (or a vote for it)

Other commands, including commands to adjust various game settings,
are still restricted. This is because most other commands would
require almost-arbitrary parameters to be useful, and so they would
provide an avenue to bypass a chat restriction.

A few commands were not whitelisted because they are more rarely
used; such commands could be whitelisted in the future if there is
a need.