Recheck lobby restrictions after certain commands are used.
Currently, changing rating/chevron restrictions for a lobby does not
affect existing lobby members until the end of the next match.

This patch changes that behavior, so that the new rating/chevron
restrictions are enforced shortly after any of the following
commands run:
- $minchevlevel
- $maxchevlevel
- $minratinglevel
- $maxratinglevel
- $setratinglevels

This is done by scheduling a `:recheck_membership` message in the
handlers for each of the above commands, making use of the
pre-existing `handle_info` in `consul_server.ex`.