Display "Smurf Link" as a restriction on user overview page.
Currently, When viewing the single-user overview page
("/teiserver/admin/user/$id"), it is not immediately obvious
whether or not the user has been flagged as a Smurf.

This is because, when no actions are active, the "Restrictions"
section contains a single "No restrictions" entry. However, while a
Smurf Link is *technically* not a restriction, it does prevent the
user from logging in.

This patch updates the user overview page's "Restrictions" section
as follows:
- "No restrictions" is only listed if there are no active actions
 AND no Smurf Link for the user.
- When the user has a Smurf Link, it is displayed in the same way
 as any other active actions.