Add pattern detection for SPADS `!unboss` command output.
Teiserver currently tracks who is the boss of a lobby by looking
for known outputs from the `!boss` SPADS command.

However, an `!unboss` command was recently added to SPADS, and
Teiserver currently does not recognize the output of that command.
As a result, whenever the `!unboss` command is used, Teiserver's
list of bosses becomes out of sync with SPADS's list.

This patch adds an additional pattern so that the output of
`!unboss` is recognized and parsed.