Isolate gen token task (#530)
* Warn if cache is down

Instead of crashing because the cache is down, only emit a warning. This
is going to lead to problem when we go multi node, but by that time, I
hope the cache usage will also be re-evaluated instead of used for
*everything*.

* gen_token task doesn't require the entire app to be started

This allow to generate a token without having to stop the app because of
a port conflict.