rename autohost into more generic bots
Although the concept of autohost remains, the general handling of bots
with regards to credentials shouldn't be tied to autohosts.
This enable having bots to access the API using bearer token in a
unified way.

This rather massive change is mostly only a search and replace
`s/autohost/bot/`. Some things are staying as they are:
* `Teiserver.Autohost` remains because it's used for tachyon and
  matchmaking
* the table `oauth_tokens` still has a column `autohost_id`, the
  migration will be done in a later commit. Migrations can be tricky and
  keeping these changes to their own commits is safer.