Refactor: Namespace the matchmaking messages
This is to prepare prepare a potential future refactor where matchmaking
could be handled in a different session module.
However, it is not that easy to split the matchmaking logic from the
rest because of things like disconnect and monitors going down. These
events could need several handlers to be processed, and that could get
messy.
So for now, simply namespace the messages