Restart tachyon system after each test
A better solution would be to restart the entire system, but when trying
Application.stop/start(:teiserver) I encountered a bunch a problems
with the db/sandbox/ranch servers and couldn't make it work.

So instead, only do that with the tachyon system, which is smaller and
self pretty well self-contained.