Bypass cache when checking for user names
Note that this query cannot use the existing index on the
`account_users` table. This shouldn't be used too often and thus impact
the performances too much.
Ideally, there would be unique index on `lower(name)` to fix that.

But why bypass the cache?
The cache is updated through a global hook processor
`Teiserver.HookServer` but this is disabled in tests. Trying to enable
it create a bunch of failure, which would need investigation.
Besides the tests, I want to reduce the reliance on caches. This is a
good start, especially because it is an attempt to avoid data collision
and consistency is important.