Rework existing combi helper
Merge it with this new combinatoric helper, and ditch the dependency.
The existing function isn't really about combinations, but finding a
team based on indices. It may get reworked further if/when the logic to
find all possible matches for given member is integrated, but it's a
rather large refactoring on a not so well tested part of the code, so
out of scope for now.