Fix a bug where the rating cache would be stale
When calling `create_or_update_rating` there's first a `get_rating`
which returns nil when the rating doesn't exist, and that would get
cached. After creation, the cache would not be updated.
Also fixed the cache invalidation for update by adding the season.

Let's ignore the fact that checking if a rating is there before creating
if not is racy, that's another bug.