Swap autohost and battle system in the supervision tree
First, battles **need** autohosts to be there to even start. This is
visible in `start_battle`. Autohosts merely send events to battle, and
later will track acks, but even then, there isn't any dependency from
autohost to battles.

This allow some logic where a battle can send messages to autohost in a
blocking manner as a response to update events.