Try out tests with synchronous teardown
This is an attempt to limit the number of error logs that happen when a
process attempt to access the DB during the async teardown. This happens
fairly often when a process restarts (like matchmaking queue server) in
the async callbacks, when the test has already completed and the sandbox
is gone.