Create a task to generate large amount of users
The output can be piped to a file, where each line is a json document
holding an access token.
The task assumes the user are already created so it'll be a bit longer
the first time, and is vastly faster on later invocation. This way we
should be fine eschewing the logic of refreshing tokens in the load
testing tool since it's easy to get a list of valid ones.