Add recon to dependencies
This helps when having to remote into a node and do some live debugging.
Since we don't have a fancy mesh network or vpn, it's easier to ssh into
the node and connect locally. And in this case, we need recon on the
node.
For more info about what recon is and how to use it, part 2 of "erlang
in anger" covers it (https://erlang-in-anger.com/ and it's free).