Live lobby kill switch (#790)
* Fix some types

The calls are ultimately doing `Enum.count` which only ever returns an
int, so match the typespec.

* Introduce a switch to disable live view of lobbies

I have some doubts it's going to fix anything, but until/if we get a
better view of why the server is melting down with high player counts,
this _may_ help. The lobby live view is showing with a high reduction
count at these times.