Track battle process in lobbies
So that when the battle terminates, the lobby can clean its state.
We may want something a bit more reactive since there can be some time
between the game end and the game process actually ending, but
regardless we need the monitoring link in place anyway in case the
battle process crash.