implement matchmaking queue versions
The version is used to guarantee a client joining a queue is aware of
the engines, games and maps that it must have.
As of this commit it is fully static though. The mechanism to force a
refetch from client when some asset change will come later. For now, the
entire matchmaking system is restarted when that happen.