Assets are static for matchmaking queues
Setup the map at queue creation (in the supervision tree), and remove
the async operation to fetch maps at init.
This simplifies the logic when it comes to testing with version because
it cannot change.
We will need to add some features to the queues so that they can
dynamically adjust if the engine/game/maps change now that we have a
mechanism to do so. But for now, it is gone.