Split internal vs tachyon start script
Up until now, the start script existed internally in the same form as
the tachyon version. It wasn't consistent with how the rest of tachyon
work and started to be annoying when it came to get back data from the
start script. For example, rating a match would require to parse the
user id back from a string.

This changes creates an internal version of the start script, and the
mapping to be done when sending it to an autohost.