Switch to regular Registry for parties
`gen_statem` does name resolution differently from `gen_server`.
The process must exist in the registry used with the via_tuple upon entering
the init function. But with Horde calling `register` followed
by `lookup` isn't guaranteed to return a result, and in this case
the start_link will fail.
So revert to the regular registry to sidestep the problem. If/when we go
multi node, we'll need to figure something out.
Retrying a few times when this happen seems to work, so that may be good
enough.