Do not query players from within mm queue (#892)
Move the logic to create members to outside the queue. Also store the
pid of sessions.
Storing the pid and avoiding querying the player registry is important
since this system is setup after the matchmaking one, and so may not be
available. This is especially likely to break when the system is
shutting down.
This also fix a bug where the queue would track the connection and not
the session.

Moving the member creation is to help with some sql sandbox test stupidity,
overall it doesn't really matter where the call is made.