Temporarily disable Credo Refactor checks
For checks with a lot of issues, I disabled them in the Credo
configuration. Those are:

- Credo.Check.Refactor.Nesting: 137 issues
- Credo.Check.Refactor.CyclomaticComplexity: 89 issues
- Credo.Check.Refactor.CondStatements: 58 issues

With this approach, it's easier to review the PR than having a bunch of
`# credo:disable-*` comments everywhere in the project.

Other checks had less issues, so I disabled them via Credo comments:

- Credo.Check.Refactor.MapJoin: 10 issues
- Credo.Check.Refactor.FilterFilter: 5 issues
- Credo.Check.Refactor.Apply: 5 issues
- Credo.Check.Refactor.FilterCount: 2 issues

We can then address those issues in other PRs. It also allows us to stop
ignoring all other Refactor checks when running Credo.