Use more granular latency buckets for spring commands (#907)
Bucket sizes are most often subsequent powers of some base number,
depending on required granularity in 1.2-4 range. Current buckets
have some large gaps at the low end, then switch to someting like
2x and then I think 1000 is too little.

In this change I'm using fibonacci numbers which is roughly ~1.6
base and have an overall nice progression.

For the latency range we are interested in it's still only 18
buckets, so it's totally fine from cardinality perspective, we can
even push it further if needed.